Revert "Refactor Vehicles module"
