+++ Postieri/Controllers/VehiclesController.cs
@@ -14,37 +14,63 @@ namespace Postieri.Controllers
     [ApiController]
     public class VehiclesController : ControllerBase
     {
-        private readonly IVehicleService _vehicleService;
+        private readonly DataContext _context;
 
-        public VehiclesController(IVehicleService vehicleService)
+        public VehiclesController(DataContext context)
         {
-            _vehicleService = vehicleService;
+            _context = context;
         }
 
-   
+        // GET: api/Vehicles
         [HttpGet]
-        public  ActionResult<List<Vehicle>> GetVehicles()
+        public async Task<ActionResult<IEnumerable<Vehicle>>> GetVehicles()
         {
-            return Ok(_vehicleService.GetVehicles());
+            if (_context.Vehicles == null)
+            {
+                return NotFound();
+            }
+            return await _context.Vehicles.ToListAsync();
         }
 
-        [HttpGet("GetAllAvailableVehiclesWithoutDefect")]    
-        public ActionResult<List<Vehicle>> GetAllAvailableVehiclesWithoutDefect()
+        [HttpGet("GetAllAvailableVehiclesWithoutDefect")]
+       
+        public async Task<ActionResult<IEnumerable<Vehicle>>> GetAllAvailableVehiclesWithoutDefect()
         {
-            return Ok(_vehicleService.GetAllAvailableVehiclesWithoutDefect());
+            if (_context.Vehicles == null)
+            {
+                return NotFound();
+            }
+            return await _context.Vehicles.Where(x => x.IsAvailable==true && x.HasDefect==false).ToListAsync();
         }
 
         [HttpGet("GetAllVehiclesWithDefect")]
-        public ActionResult<List<Vehicle>> GetAllVehiclesWithDefect()
+
+        public async Task<ActionResult<IEnumerable<Vehicle>>> GetAllVehiclesWithDefect()
         {
-            return Ok(_vehicleService.GetAllVehiclesWithDefect());
+            if (_context.Vehicles == null)
+            {
+                return NotFound();
+            }
+                return await _context.Vehicles.Where(x => x.HasDefect == true).ToListAsync();
+            
         }
 
-    
+        // GET: api/Vehicles/5
         [HttpGet("{id}")]
-        public ActionResult<Vehicle> GetVehicleById(int id)
+        public async Task<ActionResult<Vehicle>> GetVehicle(int id)
         {
-            return  Ok(_vehicleService.GetVehicleById(id));
+            if (_context.Vehicles == null)
+            {
+                return NotFound();
+            }
+            var vehicle = await _context.Vehicles.FindAsync(id);
+
+            if (vehicle == null)
+            {
+                return NotFound();
+            }
+
+            return vehicle;
         }
 
         // PUT: api/Vehicles/5
@@ -52,8 +78,29 @@ namespace Postieri.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> PutVehicle(int id, Vehicle vehicle)
         {
-            _vehicleService.PutVehicle(id, vehicle);
-            return Ok();
+            if (id != vehicle.Id)
+            {
+                return BadRequest();
+            }
+            var _vehicle = _context.Vehicles.FirstOrDefault(n => n.Id == id);
+            if (_vehicle != null)
+            {
+                _vehicle.Height = vehicle.Height;
+                _vehicle.Length = vehicle.Length;
+                _vehicle.Width = vehicle.Width;
+                _vehicle.LoadSpace = vehicle.Height * vehicle.Width * vehicle.Length;
+                _vehicle.LoadWeight = vehicle.LoadWeight;
+                _vehicle.CourierId = vehicle.CourierId;
+                _vehicle.Description = vehicle.Description;
+                _vehicle.HasDefect = vehicle.HasDefect;
+                _vehicle.IsAvailable = vehicle.IsAvailable;
+                _vehicle.Type = vehicle.Type;
+                _vehicle.PlateNumber = vehicle.PlateNumber;
+                await _context.SaveChangesAsync();
+            }
+
+
+            return NoContent();
         }
 
         // POST: api/Vehicles
@@ -61,22 +108,87 @@ namespace Postieri.Controllers
         [HttpPost]
         public async Task<ActionResult<Vehicle>> PostVehicle(Vehicle vehicle)
         {
-           _vehicleService.PostVehicle(vehicle);
-            return Ok();
+            if (_context.Vehicles == null)
+            {
+                return Problem("Entity set 'DataContext.Vehicles'  is null.");
+            }
+            var _vehicle = new Vehicle()
+            {
+                Id = vehicle.Id,
+                Height = vehicle.Height,
+                Length = vehicle.Length,
+                Width = vehicle.Width,
+                LoadSpace = vehicle.Height * vehicle.Width * vehicle.Length,
+                LoadWeight = vehicle.LoadWeight,
+                CourierId = vehicle.CourierId,
+                Description = vehicle.Description,
+                HasDefect = vehicle.HasDefect,
+                IsAvailable = vehicle.IsAvailable,
+                Type = vehicle.Type,
+                PlateNumber = vehicle.PlateNumber,
+
+            };
+            _context.Vehicles.Add(_vehicle);
+            await _context.SaveChangesAsync();
+
+            return CreatedAtAction("GetVehicle", new { id = vehicle.Id }, vehicle);
         }
 
+        // DELETE: api/Vehicles/5
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteVehicle(int id)
         {
-           _vehicleService.DeleteVehicle(id);
-            return Ok();
+            if (_context.Vehicles == null)
+            {
+                return NotFound();
+            }
+            var vehicle = await _context.Vehicles.FindAsync(id);
+            if (vehicle == null)
+            {
+                return NotFound();
+            }
+
+            _context.Vehicles.Remove(vehicle);
+            await _context.SaveChangesAsync();
+
+            return NoContent();
+        }
+
+        private bool VehicleExists(int id)
+        {
+            return (_context.Vehicles?.Any(e => e.Id == id)).GetValueOrDefault();
+        }
+
+        private bool VehicleInUseHasDefect(int id)
+        {
+            var vehicle = _context.Vehicles.Find(id);
+
+            return vehicle.HasDefect == true;
+        }
+
+        private int GetDefaultAvailableVehicle()
+        {
+            return _context.Vehicles.Where(x => x.IsAvailable == true && x.HasDefect == false).FirstOrDefault().Id;
         }
 
         [HttpPut("ChangeVehicle")]
         public async Task<IActionResult> ChangeVehicleAsync(string email)
         {
-           _vehicleService.ChangeVehicleAsync(email);
-            return Ok();
+            var courier = await _context.Couriers.FirstOrDefaultAsync(x => x.Email.ToLower().Equals(email.ToLower()));
+
+            if (courier == null)
+            {
+                return NotFound();
+            }
+            if (VehicleInUseHasDefect(courier.VehicleId))
+            {
+                courier.VehicleId = GetDefaultAvailableVehicle();
+                await _context.SaveChangesAsync();
+            }
+
+            return NoContent();
+
+
         }
     }
 }
