+++ Postieri/Program.cs
@@ -87,7 +87,6 @@ builder.Services.AddScoped<IRolesService, RolesService>();
 builder.Services.AddScoped<IOrderService, OrderService>();
 builder.Services.AddScoped<IDeliveryPriceService, DeliveryPriceService>();
 builder.Services.AddScoped<IBusinessIntegrationService, BusinessIntegrationService>();
-builder.Services.AddScoped<IVehicleService, VehicleService>();
 builder.Services.AddAutoMapper(typeof(OrderMapper).Assembly);
 builder.Services.AddControllersWithViews();
 builder.Services.AddScoped<ICourierService, CourierService>();
